Minor data structure correction.

+++ GDSA/Shared/Library/AKTravelSegment.swift
@@ -79,15 +79,15 @@ class AKTravelSegment: NSObject, NSCoding
         }
     }
     
-    func computeSpeed(let unit: UnitOfSpeed) -> Int16
+    func computeSpeed(let unit: UnitOfSpeed) -> Int32
     {
         switch unit {
         case UnitOfSpeed.MetersPerSecond:
-            return Int16(self.speed)
+            return Int32(self.speed)
         case UnitOfSpeed.KilometersPerHour:
-            return Int16(self.speed * 3.6)
+            return Int32(self.speed * 3.6)
         case UnitOfSpeed.MilesPerHour:
-            return Int16(self.speed * 2.23694)
+            return Int32(self.speed * 2.23694)
         }
     }
